# chore: optimize CollectionInfoBuilder